requirements.txt
Flask==3.0.0
requests==2.31.0
pycryptodome==3.19.0
protobuf==4.25.1
gunicorn==21.2.0